separação da visão do usuario e do restaurante
